+++ packages/server/routes/infoarch.js
@@ -26,6 +26,7 @@ const Snapshot = require("@saltcorn/admin-models/models/snapshot");
 const { stringify } = require("csv-stringify");
 const csvtojson = require("csvtojson");
 const { hasLLM, translate } = require("@saltcorn/data/translate");
+const { escapeHtml } = require("@saltcorn/data/utils");
 
 /**
  * @type {object}
@@ -365,7 +366,7 @@ router.get(
                 [
                   {
                     label: req.__("In default language"),
-                    key: "in_default",
+                    key: (r) => escapeHtml(r.in_default),
                   },
                   {
                     label: req.__("In %s", form.values.name),
@@ -375,8 +376,9 @@ router.get(
                           "data-inline-edit-dest-url": `/site-structure/localizer/save-string/${lang}/${encodeURIComponent(
                             r.in_default
                           )}`,
+                          "data-inline-edit-unescape": "true",
                         },
-                        r.translated
+                        escapeHtml(r.translated)
                       ),
                   },
                 ],
@@ -431,10 +433,12 @@ router.post(
       req.flash(
         "success",
         req.__(
-          `Translated %s strings. Click 'Translate with LLM' again to continue`, count
+          `Translated %s strings. Click 'Translate with LLM' again to continue`,
+          count
         )
       );
-    else req.flash("success", req.__(`Finished translating %s strings.`, count));
+    else
+      req.flash("success", req.__(`Finished translating %s strings.`, count));
     res.redirect(`/site-structure/localizer/edit/${lang}`);
   })
 );
@@ -459,9 +463,8 @@ router.post(
       return;
     }
     const cfgStrings = getState().getConfigCopy("localizer_strings");
-    if (cfgStrings[lang])
-      cfgStrings[lang][defstring] = text((req.body || {}).value);
-    else cfgStrings[lang] = { [defstring]: text((req.body || {}).value) };
+    if (cfgStrings[lang]) cfgStrings[lang][defstring] = (req.body || {}).value;
+    else cfgStrings[lang] = { [defstring]: (req.body || {}).value };
     await getState().setConfig("localizer_strings", cfgStrings);
     res.redirect(`/site-structure/localizer/edit/${lang}`);
   })